move sendUserCode method from helpers to users
